requirements.txt
openai
python-dotenv
gradio
redis
redis[hiredis]